+++ modules/tinymce/src/core/test/ts/browser/undo/ForcedRootBlockTest.ts
@@ -4,7 +4,6 @@ import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
 import * as Levels from 'tinymce/core/undo/Levels';
-import { UndoLevelType } from 'tinymce/core/undo/UndoManagerTypes';
 
 describe('browser.tinymce.core.undo.ForcedRootBlockTest', () => {
   const hook = TinyHooks.bddSetupLight<Editor>({
@@ -24,7 +23,7 @@ describe('browser.tinymce.core.undo.ForcedRootBlockTest', () => {
       bookmark: null,
       content: '<strong>a</strong> <span>b</span>',
       fragments: null,
-      type: UndoLevelType.Complete
+      type: 'complete'
     });
   });
 
@@ -43,7 +42,7 @@ describe('browser.tinymce.core.undo.ForcedRootBlockTest', () => {
         ' ',
         '<span>b</span>'
       ],
-      type: UndoLevelType.Fragmented
+      type: 'fragmented'
     });
   });
 });